requirements.txt
sionna==0.19.0
mitsuba==3.4.0
tensorflow[and-cuda]
jupyter
hydra-core
tinydb
runipy
pandas
gymnasium